+++ MangaRipper.Plugin.MangaFox/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Selenium.WebDriver" version="3.141.0" targetFramework="net462" />
+</packages>
